Add dev/prod profiles, document code, rename packages
