+++ ipb-platform/src/main/java/com/ipb/db/entity/IpbImage.java
@@ -1,4 +1,4 @@
-package com.ipb.entity;
+package com.ipb.db.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -8,6 +8,18 @@ import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
+/**
+ * This class represents 
+ * an "image"
+ * in the database.
+ * 
+ * It is converted to a table in the database via Hibernate.
+ * The fields in the class become the table's attributes. 
+ * Certain restrictions on the attributes' length/value are set via JPA annotations
+ * and are then applied by Hibernate when creating the tables.
+ * 
+ * @author Dimitar Trifonov (dvt32)
+ */
 @Entity
 @Table(name = "Images")
 public class IpbImage {
