+++ ipb-platform/src/main/java/com/ipb/db/entity/IpbRouteObject.java
@@ -1,4 +1,4 @@
-package com.ipb.entity;
+package com.ipb.db.entity;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -6,10 +6,25 @@ import javax.persistence.Id;
 import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 
+/**
+ * This class represents 
+ * the relationship between "routes" and "objects"
+ * in the database.
+ * 
+ * It is converted to a table in the database via Hibernate.
+ * The fields in the class become the table's attributes. 
+ * Certain restrictions on the attributes' length/value are set via JPA annotations
+ * and are then applied by Hibernate when creating the tables.
+ * 
+ * NOTE: This class may become unnecessary when replaced 
+ * by proper JPA annotations for establishing table relationships.
+ * 
+ * @author Dimitar Trifonov (dvt32)
+ */
 @Entity
-@Table(name = "Categories_Objects")
-public class IpbCategoryObject {
-
+@Table(name = "Routes_Objects")
+public class IpbRouteObject {
+	
 	@Id
 	@GeneratedValue
 	private Long id;
@@ -23,7 +38,7 @@ public class IpbCategoryObject {
 	/*
 	 * Constructors
 	 */
-	public IpbCategoryObject() {
+	public IpbRouteObject() {
 		
 	}
