--- ipb-platform/src/main/java/com/ipb/repository/IpbCategoryRepository.java
@@ -1,9 +1,0 @@
-package com.ipb.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ipb.entity.IpbCategory;
-
-public interface IpbCategoryRepository extends JpaRepository<IpbCategory, Long> {
-
-}
