--- ipb-platform/src/main/java/com/ipb/repository/IpbEventRepository.java
@@ -1,9 +1,0 @@
-package com.ipb.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ipb.entity.IpbEvent;
-
-public interface IpbEventRepository extends JpaRepository<IpbEvent, Long> {
-	
-}
