--- ipb-platform/src/main/java/com/ipb/repository/IpbImageRepository.java
@@ -1,9 +1,0 @@
-package com.ipb.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ipb.entity.IpbImage;
-
-public interface IpbImageRepository extends JpaRepository<IpbImage, Long> {
-
-}
