--- ipb-platform/src/main/java/com/ipb/repository/IpbLandmarkRepository.java
@@ -1,9 +1,0 @@
-package com.ipb.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ipb.entity.IpbLandmark;
-
-public interface IpbLandmarkRepository extends JpaRepository<IpbLandmark, Long> {
-
-}
