--- ipb-platform/src/main/java/com/ipb/repository/IpbObjectRepository.java
@@ -1,9 +1,0 @@
-package com.ipb.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ipb.entity.IpbObject;
-
-public interface IpbObjectRepository extends JpaRepository<IpbObject, Long> {
-
-}
