--- ipb-platform/src/main/java/com/ipb/repository/IpbProposalFileRepository.java
@@ -1,9 +1,0 @@
-package com.ipb.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ipb.entity.IpbProposalFile;
-
-public interface IpbProposalFileRepository extends JpaRepository<IpbProposalFile, Long> {
-
-}
