--- ipb-platform/src/main/java/com/ipb/repository/IpbProposalRepository.java
@@ -1,9 +1,0 @@
-package com.ipb.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ipb.entity.IpbProposal;
-
-public interface IpbProposalRepository extends JpaRepository<IpbProposal, Long> {
-
-}
