--- ipb-platform/src/main/java/com/ipb/repository/IpbRouteRepository.java
@@ -1,9 +1,0 @@
-package com.ipb.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ipb.entity.IpbRoute;
-
-public interface IpbRouteRepository extends JpaRepository<IpbRoute, Long> {
-
-}
