--- ipb-platform/src/main/java/com/ipb/repository/IpbUserRepository.java
@@ -1,9 +1,0 @@
-package com.ipb.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ipb.entity.IpbUser;
-
-public interface IpbUserRepository extends JpaRepository<IpbUser, Long> {
-
-}
