+++ ipb-platform/src/main/resources/application-dev.properties
@@ -1,0 +1,10 @@
+# Database settings
+spring.jpa.hibernate.ddl-auto = create
+spring.datasource.url = jdbc:oracle:thin:@localhost:1521:orcl
+spring.datasource.driver-class-name = oracle.jdbc.OracleDriver
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.Oracle12cDialect
+spring.jpa.show-sql = true
+
+# Replace username/login with your own local username/password
+spring.datasource.username = c##dvt32
+spring.datasource.password = 123
