+++ ipb-platform/src/main/resources/application.properties
@@ -1,8 +1,2 @@
-# Database settings
-spring.jpa.hibernate.ddl-auto = create
-spring.datasource.url = jdbc:oracle:thin:@localhost:1521:orcl
-spring.datasource.username = c##dvt32
-spring.datasource.password = 123
-spring.datasource.driver-class-name = oracle.jdbc.OracleDriver
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.Oracle12cDialect
-spring.jpa.show-sql = true
+spring.profiles.active = dev
+spring.application.name = ipb-platform
